IDEX-3226: fix docker container start in DockerInstanceProvider
In some cases we need add extra host records to hosts file of
container. Variable that provides extra hosts is array of strings.
If this variable is bound in property file and value of property
is empty array with 1 empty string is injected by Guice.
It is illegal exta hosts value for docker.
I commented usage of extra hosts and tests for extra hosts
addition. We need fix issue with illegal value of the array to
uncomment its usage.